Refactor user service imports and split user service functions into separate files
