+++ src/functions/user/services/availability/generate.ts
@@ -8,7 +8,8 @@ import { findStartAndEndDate } from "~/library/availability/find-start-end-date-
 import { generateAvailability } from "~/library/availability/generate-availability";
 import { removeBookedSlots } from "~/library/availability/remove-booked-slots";
 import { StringOrObjectId } from "~/library/zod";
-import { UserServiceGetCustomerId } from "../user";
+
+import { UserServiceGetCustomerId } from "../user/get-customer-id";
 import { UserAvailabilityServiceGetOrders } from "./get-orders";
 
 export type UserAvailabilityServiceGenerateProps = {
