+++ src/functions/user/services/user/get.ts
@@ -1,0 +1,19 @@
+import { NotFoundError } from "~/library/handler";
+import { UserModel } from "../../user.model";
+import { User } from "../../user.types";
+
+export type UserServiceGetProps = Required<Pick<User, "username">>;
+
+export const UserServiceGet = ({ username }: UserServiceGetProps) => {
+  return UserModel.findOne({ username, active: true, isBusiness: true })
+    .lean()
+    .orFail(
+      new NotFoundError([
+        {
+          path: ["username"],
+          message: "NOT_FOUND",
+          code: "custom",
+        },
+      ])
+    );
+};
