+++ src/functions/user/services/user/username-taken.spec.ts
@@ -1,0 +1,25 @@
+import { CustomerServiceCreate } from "~/functions/customer/services/customer";
+import { getUserObject } from "~/library/jest/helpers";
+import { UserServiceUsernameTaken } from "./username-taken";
+
+require("~/library/jest/mongoose/mongodb.jest");
+
+describe("UserServiceUsernameTaken", () => {
+  it("check username is taken", async () => {
+    const userData = getUserObject();
+
+    await CustomerServiceCreate(userData);
+
+    let response = await UserServiceUsernameTaken({
+      username: userData.username,
+    });
+
+    expect(response.usernameTaken).toBeTruthy();
+
+    response = await UserServiceUsernameTaken({
+      username: "testerne",
+    });
+
+    expect(response.usernameTaken).toBeFalsy();
+  });
+});
